+++ CHANGELOG.md
@@ -21,6 +21,7 @@ All notable changes to this project will be documented in this file.
 #### Supabase
 
 - add user roles table migration
+- add user role trigger migration
 
 ### Changed